+++ src/lib/components/PositionDisplay.svelte
